# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Mangaty API
  description: |
    API para la plataforma de cómics Mangaty.
    
    ## Características:
    - Autenticación segura con JWT
    - Sistema de roles flexible (lector → autor → admin)
    - Microeconomía con TyCoins
    - Subida de imágenes con Cloudinary
    - Sistema de recomendaciones con IA
    
    ## Roles del sistema:
    - **lector**: Rol por defecto. Lee cómics y desbloquea capítulos
    - **autor**: Puede crear cómics y publicar capítulos. Cualquier lector puede subir contenido
    - **administrador**: Control total. Asigna roles y elimina contenido
    
    ## Flujo de autenticación:
    1. Registrar usuario → `POST /api/v1/auth/register`
    2. Iniciar sesión → `POST /api/v1/auth/login`
    3. Clic en **Authorize** y pegar: `Bearer <token>`
  version: 1.0.0
  contact:
    name: Mangaty Team
    email: support@mangaty.com
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://api.mangaty.com
    description: Producción
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT del login. Formato: `Bearer <token>`
        El token incluye los roles del usuario
  schemas:
...
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time

    # ============ AUTH ============
...
    WalletData:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        balance:
          type: integer
          minimum: 0
        currency:
          type: string
          example: TyCoins
    UnlockChapterRequest:
      type: object
      required: [chapter_id]
      properties:
        chapter_id:
          type: string
          format: uuid

    # ============ COMICS ============
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
...
    BadRequest:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/v1/wallet/balance:
    get:
      tags: [Wallet]
      summary: Consultar saldo
      operationId: getBalance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saldo en TyCoins
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WalletData'
  /api/v1/wallet/unlock-chapter:
    post:
      tags: [Wallet]
      summary: Desbloquear capítulo
      description: Paga con TyCoins para desbloquear un capítulo
      operationId: unlockChapter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockChapterRequest'
      responses:
        '200':
          description: Capítulo desbloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: Saldo insuficiente
          $ref: '#/components/responses/BadRequest'

  # ============ COMICS ============
...
security: []